Rework ingame hud struct (#442)
* name some camera things

* yea there

* lots of audio

* match a func

* better

* x

* egg

* all but one func in audio.c nopw

* audio looks good for now

* Update audio.c

* correct taj thing

* fix warnings

* Update audiosfx.h

* Update README.md

* preliminary names

* fix warnings

* fix an oopsie

* checkpoint

* oh my god

* Update README.md

* run clang tidy

* initial hud asset ids

+++ src/menu.c
@@ -8886,7 +8886,7 @@ void render_track_select_setup_ui(UNUSED s32 updateRate) {
             sMenuGuiOpacity = 0;
         }
         availableVehicleFlags = get_map_available_vehicles(gTrackIdForPreview);
-        i = get_level_name(gTrackIdForPreview);
+        i = (s32) get_level_name(gTrackIdForPreview);
         set_text_font(ASSET_FONTS_BIGFONT);
         set_text_colour(192, 192, 255, 0, sMenuGuiOpacity);
         set_text_background_colour(0, 0, 0, 0);
